+++ 掌上云购/掌上云购/MainCode/Person/Controller/CustomerCare /CustomerCareController.m
@@ -124,6 +124,6 @@
         default:
             break;
     }
-    NSLogZS(@"%ld",sender.selectedSegmentIndex);
+    NSLogZS(@"%ld",(long)sender.selectedSegmentIndex);
 }
 @end
